B100788 Enforcing the string parameter in prStrWidth
Without this, Perl throws warnings to the effect of "Use of
uninitialized value $string in upack.." We really need a much
more robust system of error trapping and communication back
to the caller.

Thanks to Mark Tompsett for pointing me to the deb fix.
Thanks also to Patrick Matthäi for patching this upstream.