B97290 Setting Index parameter to its standard default
When trying to create a new PDF file using "prDoc" by extracting a
few pages from an existing PDF version 1.5 file, results in this
error message: Use of uninitialized value $param{"Index"} in pattern
match (m//) at blib/lib/PDF/Reuse.pm
(autosplit into blib/lib/auto/PDF/Reuse/crossrefObj.al) line 4275.
as well as various other error messages. According to the PDF
standard, the "Index" parameter is optional, and has a default value
of "[0 Size]" This patch adds a check to the crossrefObj subroutine
which sets this parameter to the default value if it is missing.

Thanks to Ken Neighbors for this fix.