B94624 Word/Character spacing is incorrect after transformations
Text printed with a rotation has too much word spacing. This appears
to be caused by the failure to set Tw and Tc to at least a default
value of zero. This patch sets these parameters to zero. A later fix
is needed to allow these parameters to be passed in by the caller.

Thanks to Jarom Smith for catching this bug.