B46202 PDF-Reuse 0.35 prDoc/xrefSection problem
PDF::Reuse crashes using prDoc with some pdfs created with Xerox X-Solutions
ScanFlowStore

Solution: change xrefSection move $buf contatination to beginning of while loop.
It looks like xrefSection is expecting a blank line which it can safely ignore
to follow the xref section before the trailer line, but the ScanFlowStore
software does not always include the blank line. By moving the '$buf .=' line,
the line used to break out of the xref processing is scanned for the trailer information.

Thanks to Steve Meister for this fix.