B120401 Reads the entire contents when extracting a single page
Both prDoc and prSinglePage read the entire contents of the source PDF,
even when extracting just a single page. Extracting a page from a 160MB
PDF with: prFile("pr.pdf"); prSinglePage('source.pdf', 1); prEnd();
produced a 160MB file with one page.

PDF::Reuse assumes that the word ‘xref’, followed by a return, occurs at
that location, so it skips 5 bytes before reading the xref entries. So
for the PDF in question it was starting at the ‘f’ in xref and, not seeing
any numbers on that line, assumed it had reached the end. Since the PDF
was linearized and had two cross-reference tables, it was only reading the
first one (with about 50 entries), whereas the combined cross-reference
table was supposed to have nearly two thousand entries.

The way PDF::Reuse reads objects is interesting. It sorts all the offsets
in the xref table and then uses them to determine how many bytes to read
for each object, under the assumption that the next offset marks the end
of the object. (Makes sense. It also makes PDF::Reuse immune to bugs like
B120397.) if the xref table was misread, then the objects whose xref
entries were not read will be considered part of whichever object preceded
them in the file. So *that* explains the file bloat.

This patch fixes the problem, providing a major performance increase for
page extraction. PDF::Reuse will extract a page from a 169.5 MB file in a
split second, rather than 15 seconds.

Thanks to Father Chrysostomos (SPROUT) for finding and fixing this!